+++ 3750/2/path_generator.rb
@@ -1,13 +1,22 @@
-module PathGenerator
+class PathGenerator
   TOKEN = ENV['BOT_TOKEN']
 
   BOT_API_URL = "https://api.telegram.org/bot#{TOKEN}/".freeze
   BOT_DOWNLOAD_API_URL = "https://api.telegram.org/file/bot#{TOKEN}/".freeze
   GET_PATH_URL = 'getFile?file_id='.freeze
 
-  def save_path(timestamp)
-    return checkin_path(timestamp) if session[:command] == 'checkin'
-    checkout_path(timestamp)
+  def initialize(session, payload)
+    @timestamp = session[:timestamp]
+    @command = session[:command]
+    @id = session[:id]
+    @payload = payload
+  end
+
+  attr_reader :timestamp, :command, :id, :payload
+
+  def save_path
+    return checkin_path if command == 'checkin'
+    checkout_path
   end
 
   def download_path
@@ -25,12 +34,12 @@ module PathGenerator
     payload.fetch('photo', [{}]).last.fetch('file_id')
   end
 
-  def checkin_path(timestamp)
-    "./public/#{from['id']}/checkins/#{timestamp}/"
+  def checkin_path
+    "./public/#{id}/checkins/#{timestamp}/"
   end
 
-  def checkout_path(timestamp)
-    "./public/#{from['id']}/checkouts/#{timestamp}/"
+  def checkout_path
+    "./public/#{id}/checkouts/#{timestamp}/"
   end
 
   def path_request_url
