--- 3750/2/validators/photo_validator.rb
@@ -1,5 +1,0 @@
-module PhotoValidator
-  def photo?
-    payload['photo']
-  end
-end
